fixed, let's try again at fixing this
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/trunk@29265 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/VideoDatabase.cpp
@@ -5046,7 +5046,7 @@ bool CVideoDatabase::GetTvShowsByWhere(const CStdString& strBaseDir, const CStdS
               CTimeUtils::GetTimeMS() - time);
 
     CStdString order(where);
-    bool maintainOrder = (size_t)order.ToLower().Find("order by") != -1;
+    bool maintainOrder = order.ToLower().Find("order by") != -1;
     Stack(items, VIDEODB_CONTENT_TVSHOWS, maintainOrder);
 
     // cleanup
